+++ CHANGELOG.md
@@ -9,3 +9,4 @@ Every sprint's progress reflected here.
 ### Added
 
 - Registration Page [#29](https://github.com/mulla028/T9-Final_Project/pull/29)
+- Forgot Password logic and UI/UX using `nodemailer` Added [#28](https://github.com/mulla028/T9-Final_Project/pull/28)